requirements.txt
requests
python-dotenv
pandas
web3
solders
solana
